Moving buildCommand into module commandBuilder
