+++ tasks/mspec.js
@@ -9,45 +9,9 @@
 'use strict';
 
 var path = require('path'),
-  buildCommand = function(grunt, files, options) {
-    var mspec = options.platform === 'x86' ? 'mspec-x86-clr4.exe' : 'mspec-clr4.exe';
-    if (options.toolsPath) {
-      if (!grunt.file.isPathAbsolute(options.toolsPath)) {
-        options.toolsPath = path.join(process.cwd(), options.toolsPath);
-      }
-      mspec = path.join(options.toolsPath, mspec);
-    }
-    mspec = mspec.replace(/\\/g, path.sep);
-    var assemblies = files.map(function(file) {
-      return '"' + file.src + '"';
-    });
-    var args = assemblies;
-
-    if (options.timeinfo) {
-      args.unshift('-t');
-    }
-    if (options.silent) {
-      args.unshift('-s');
-    }
-    if (options.progress) {
-      args.unshift('-p');
-    }
-    if (options.output) {
-      var filePath = path.join(process.cwd(), options.output);
-      grunt.file.mkdir(filePath);
-      args.unshift('"' + path.join(filePath, 'index.xml') + '"');
-      args.unshift('--xml');
-      args.unshift('"' + filePath + '"');
-      args.unshift('--html');
-    }
-
-    return {
-      path: path.normalize(mspec),
-      args: args
-    };
-  };
-var name = 'mspec';
-var description = 'Grunt plugin to run Machine.Specfication tests';
+    buildCommand = require('./mspec.commandBuilder'),
+    name = 'mspec',
+    description = 'Grunt plugin to run Machine.Specfication tests';
 
 module.exports = function(grunt) {
 
@@ -62,7 +26,7 @@ module.exports = function(grunt) {
     }),
       files = this.files,
       taskComplete = this.async();
-    var command = buildCommand(grunt, files, options);
+    var command = new buildCommand(grunt, files, options);
 
     console.log();
     console.log('mspec test runner');
